requirements.txt
anyio==3.6.2
aptos-sdk==0.5.1
certifi==2022.12.7
cffi==1.15.1
grpcio==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
mypy==1.2.0
mypy-extensions==1.0.0
protobuf==4.22.3
pycparser==2.21
PyNaCl==1.5.0
rfc3986==2.0.0
sniffio==1.3.0
typing_extensions==4.5.0
